# api/apidocs/apiDoc.v1.0.0.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "1.0.0",
    "title": "Lew & Lowski APIs Document",
    "description": "Nodejs Based API",
    "termsOfService": "",
    "contact": {
      "name": "Parveen Chauhan",
      "email": "pkumar@lewanddowski.com",
      "url": "https://lewanddowski.com"
    }
  },
  "servers": [
    {
      "url": "/api/v1.0.0/",
      "description": "API verion"
    }
  ],
...
  "paths": {
...
    "/logitix/getEventByProductionId": {
      "post": {
        "tags": ["Logitix"],
        "description": "Get Event By Production Id",
        "operationId": "getEventByProductionId",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": ["productionId"],
                "properties": {
                  "productionId": {
                    "type": "string",
                    "description": "Event Production Id "
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Response"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "jwt": {
        "name": "x-access-token",
        "type": "Api Key",
        "scheme": "bearer",
        "in": "header",
        "bearerFormat": "Access Token"
      }
    }
  }
}